feat: support main function arguments in scripts (#620)
This PR closes #591 by making it easy to pass arguments to the `main`
function of a Sway script.

* Adds a `script_abigen` macro that generates the Rust binding with the
script types.
* Completely reworks the way scripts and calls are handled.
