+++ docs/src/SUMMARY.md
@@ -35,6 +35,7 @@
     - [Calling other contracts](./calling-contracts/other-contracts.md)
     - [Multiple contract calls](./calling-contracts/multicalls.md)
     - [Estimating cost](./calling-contracts/cost-estimation.md)
+  - [Running scripts](./getting-started/running-scripts.md)
   - [Predicates](./getting-started/predicates.md)
     - [Send and spend funds](./predicates/send-spend-predicate.md)
   - [Scripts](./scripts/index.md)
