+++ docs/src/calling-contracts/multicalls.md
@@ -24,8 +24,8 @@ To get the output values of the bundled calls, you need to provide explicit type
 {{#include ../../../examples/contracts/src/lib.rs:multi_call_values}}
 ```
 
-You can also interact with the `CallResponse` by moving the type annotation to the invoked method:
+You can also interact with the `FuelCallResponse` by moving the type annotation to the invoked method:
 
 ```rust,ignore
-{{#include ../../../examples/contracts/src/lib.rs:multi_call_response}}
+{{#include ../../../examples/contracts/src/lib.rs:multi_contract_call_response}}
 ```
