+++ docs/src/testing/index.md
@@ -1,17 +1,5 @@
 # `fuels-rs` Testing
 
-> **note** This page is still a work in progress.
-
-## Running a Sway script
-
-`run_compiled_script` is a helper function for testing simple Sway scripts and reducing boilerplate code related to setting up contracts and deployment. It takes the path to the generated `.bin` file as an argument.
-
-- You can use it this way:
-
-````rust,ignore
-{{#include ../../../packages/fuels-test-helpers/src/script.rs:run_compiled_script}}
-````
-
-Here you'll find an overview of the various test helpers located in `fuels-test-helpers`:
+**Note** This page is still a work in progress
 
 - [Tweaking the blockchain](./chains.md)
