+++ packages/fuels-contract/src/lib.rs
@@ -1,7 +1,10 @@
+pub mod call_response;
 pub mod contract;
+pub mod contract_calls_utils;
+pub mod execution_script;
 pub mod logs;
 pub mod predicate;
-pub mod script;
+pub mod script_calls;
 
 pub mod abi_encoder {
     pub use fuels_core::abi_encoder::*;
