+++ packages/fuels-core/src/code_gen/abigen.rs
@@ -3,10 +3,12 @@ use super::{
     functions_gen::expand_function,
     resolved_type::resolve_type,
 };
-use crate::{code_gen::bindings::ContractBindings, source::Source, utils::ident};
+use crate::code_gen::{bindings::ContractBindings, functions_gen::generate_script_main_function};
+use crate::source::Source;
+use crate::utils::ident;
 use fuels_types::{
     bech32::Bech32ContractId, errors::Error, param_types::ParamType, utils::custom_type_name,
-    ProgramABI, ResolvedLog, TypeDeclaration,
+    ABIFunction, ProgramABI, ResolvedLog, TypeDeclaration,
 };
 use inflector::Inflector;
 use proc_macro2::TokenStream;
@@ -18,8 +20,11 @@ pub struct Abigen {
     rustfmt: bool,
     /// Generate no-std safe code
     no_std: bool,
-    contract_name: String,
+    /// The contract or script name as an identifier.
+    name: String,
+
     abi: ProgramABI,
+
     types: HashMap<usize, TypeDeclaration>,
 }
 
@@ -34,7 +39,7 @@ impl Abigen {
         Ok(Self {
             types: Abigen::get_types(&parsed_abi),
             abi: parsed_abi,
-            contract_name: contract_name.to_string(),
+            name: contract_name.to_string(),
             rustfmt: true,
             no_std: false,
         })
@@ -48,120 +53,86 @@ impl Abigen {
     /// Generates the contract bindings.
     pub fn generate(self) -> Result<ContractBindings, Error> {
         let rustfmt = self.rustfmt;
-        let tokens = self.expand()?;
+        let tokens = self.expand_contract()?;
 
         Ok(ContractBindings { tokens, rustfmt })
     }
 
     /// Entry point of the Abigen's expansion logic.
-    /// The high-level goal of this function is to expand* a contract
-    /// defined as a JSON into type-safe bindings of that contract that can be
-    /// used after it is brought into scope after a successful generation.
+    /// The high-level goal of this function is to expand* a contract defined as a JSON ABI
+    /// into type-safe bindings of that contract that can be used after it is brought into
+    /// scope after a successful generation.
     ///
-    /// *: To expand, in procedural macro terms, means to automatically generate
-    /// Rust code after a transformation of `TokenStream` to another
-    /// set of `TokenStream`. This generated Rust code is the brought into scope
-    /// after it is called through a procedural macro (`abigen!()` in our case).
-    pub fn expand(&self) -> Result<TokenStream, Error> {
-        let name = ident(&self.contract_name);
+    /// *: To expand, in procedural macro terms, means to automatically generate Rust code after a
+    /// transformation of `TokenStream` to another set of `TokenStream`. This generated Rust code is
+    /// the brought into scope after it is called through a procedural macro
+    /// (`abigen!()` in our case).
+    pub fn expand_contract(&self) -> Result<TokenStream, Error> {
+        let name = ident(&self.name);
         let methods_name = ident(&format!("{}Methods", name));
-        let name_mod = ident(&format!(
-            "{}_mod",
-            self.contract_name.to_string().to_snake_case()
-        ));
+        let name_mod = ident(&format!("{}_mod", self.name.to_string().to_snake_case()));
 
-        let contract_functions = self.functions()?;
+        let contract_functions = self.contract_functions()?;
         let abi_structs = self.abi_structs()?;
         let abi_enums = self.abi_enums()?;
 
         let resolved_logs = self.resolve_logs();
         let log_id_param_type_pairs = generate_log_id_param_type_pairs(&resolved_logs);
 
-        let (includes, code) = if self.no_std {
-            (
-                quote! {
-                    use alloc::{vec, vec::Vec};
-                    use fuels_core::{EnumSelector, Parameterize, Tokenizable, Token, Identity, try_from_bytes};
-                    use fuels_core::types::*;
-                    use fuels_core::code_gen::function_selector::resolve_fn_selector;
-                    use fuels_types::errors::Error as SDKError;
-                    use fuels_types::param_types::ParamType;
-                    use fuels_types::enum_variants::EnumVariants;
-                },
-                quote! {},
-            )
+        let includes = self.includes(false);
+
+        let code = if self.no_std {
+            quote! {}
         } else {
-            (
-                quote! {
-                    use fuels::contract::{contract::{Contract, ContractCallHandler}, logs::LogDecoder};
-                     use fuels::core::{EnumSelector, StringToken, Parameterize, Tokenizable, Token,
-                                      Identity, try_from_bytes};
-                    use fuels::core::code_gen::get_logs_hashmap;
-                    use fuels::core::abi_decoder::ABIDecoder;
-                    use fuels::core::code_gen::function_selector::resolve_fn_selector;
-                    use fuels::core::types::*;
-                    use fuels::signers::WalletUnlocked;
-                    use fuels::tx::{ContractId, Address, Receipt};
-                    use fuels::types::bech32::Bech32ContractId;
-                    use fuels::types::ResolvedLog;
-                    use fuels::types::errors::Error as SDKError;
-                    use fuels::types::param_types::ParamType;
-                    use fuels::types::enum_variants::EnumVariants;
-                    use std::str::FromStr;
-                    use std::collections::{HashSet, HashMap};
-                    use std::fmt;
-                },
-                quote! {
-                    pub struct #name {
-                        contract_id: Bech32ContractId,
-                        wallet: WalletUnlocked,
+            quote! {
+                pub struct #name {
+                 contract_id: Bech32ContractId,
+                 wallet: WalletUnlocked,
+                }
+
+                impl #name {
+                    pub fn new(contract_id: Bech32ContractId, wallet: WalletUnlocked) -> Self {
+                        Self { contract_id, wallet}
                     }
 
-                    impl #name {
-                        pub fn new(contract_id: Bech32ContractId, wallet: WalletUnlocked) -> Self {
-                            Self { contract_id, wallet}
-                        }
+                    pub fn get_contract_id(&self) -> &Bech32ContractId {
+                         &self.contract_id
+                     }
 
-                        pub fn get_contract_id(&self) -> &Bech32ContractId {
-                            &self.contract_id
-                        }
+                    pub fn get_wallet(&self) -> WalletUnlocked {
+                         self.wallet.clone()
+                     }
 
-                        pub fn get_wallet(&self) -> WalletUnlocked {
-                            self.wallet.clone()
-                        }
+                    pub fn with_wallet(&self, mut wallet: WalletUnlocked) -> Result<Self, SDKError> {
+                        let provider = self.wallet.get_provider()?;
+                        wallet.set_provider(provider.clone());
+                        Ok(Self { contract_id: self.contract_id.clone(), wallet: wallet})
+                     }
 
-                        pub fn with_wallet(&self, mut wallet: WalletUnlocked) -> Result<Self, SDKError> {
-                           let provider = self.wallet.get_provider()?;
-                           wallet.set_provider(provider.clone());
-
-                           Ok(Self { contract_id: self.contract_id.clone(), wallet: wallet})
-                        }
-
-                        pub async fn get_balances(&self) -> Result<HashMap<String, u64>, SDKError> {
-                            self.wallet.get_provider()?.get_contract_balances(&self.contract_id).await.map_err(Into::into)
-                        }
-
-                        pub fn methods(&self) -> #methods_name {
-                            #methods_name {
-                                contract_id: self.contract_id.clone(),
-                                wallet: self.wallet.clone(),
-                                logs_map: get_logs_hashmap(&[#(#log_id_param_type_pairs),*], &self.contract_id),
-                            }
-                        }
+                    pub async fn get_balances(&self) -> Result<HashMap<String, u64>, SDKError> {
+                        self.wallet.get_provider()?.get_contract_balances(&self.contract_id).await.map_err(Into::into)
                     }
 
-                    // Implement struct that holds the contract methods
-                    pub struct #methods_name {
-                        contract_id: Bech32ContractId,
-                        wallet: WalletUnlocked,
-                        logs_map: HashMap<(Bech32ContractId, u64), ParamType>,
+                    pub fn methods(&self) -> #methods_name {
+                        #methods_name {
+                            contract_id: self.contract_id.clone(),
+                            wallet: self.wallet.clone(),
+                            logs_map: get_logs_hashmap(&[#(#log_id_param_type_pairs),*], &self.contract_id),
+                        }
                     }
+                }
 
-                    impl #methods_name {
-                        #contract_functions
-                    }
-                },
-            )
+                // Implement struct that holds the contract methods
+                pub struct #methods_name {
+                    contract_id: Bech32ContractId,
+                    wallet: WalletUnlocked,
+                    logs_map: HashMap<(Bech32ContractId, u64), ParamType>,
+                }
+
+                impl #methods_name {
+                    #contract_functions
+                }
+            }
         };
 
         Ok(quote! {
@@ -184,15 +155,137 @@ impl Abigen {
         })
     }
 
-    pub fn functions(&self) -> Result<TokenStream, Error> {
+    /// Expand a script into type-safe Rust bindings based on its ABI. See `expand_contract` for
+    /// more details.
+    pub fn expand_script(&self) -> Result<TokenStream, Error> {
+        let name = ident(&self.name);
+        let name_mod = ident(&format!("{}_mod", self.name.to_string().to_snake_case()));
+
+        let includes = self.includes(true);
+
+        let main_script_function = self.script_function()?;
+        let code = if self.no_std {
+            quote! {}
+        } else {
+            quote! {
+                #[derive(Debug)]
+                pub struct #name{
+                    wallet: WalletUnlocked,
+                    binary_filepath: String,
+                }
+
+                impl #name {
+                    pub fn new(wallet: WalletUnlocked, binary_filepath: &str) -> Self {
+                        Self {wallet: wallet, binary_filepath: binary_filepath.to_string()}
+                    }
+
+                    #main_script_function
+                }
+            }
+        };
+
+        let abi_structs = self.abi_structs()?;
+        let abi_enums = self.abi_enums()?;
+        Ok(quote! {
+            pub use #name_mod::*;
+
+            #[allow(clippy::too_many_arguments)]
+            pub mod #name_mod {
+                #![allow(clippy::enum_variant_names)]
+                #![allow(dead_code)]
+
+                #includes
+
+                #code
+
+                #abi_structs
+                #abi_enums
+
+            }
+        })
+    }
+
+    /// Generates the includes necessary for the abigen.
+    fn includes(&self, is_script: bool) -> TokenStream {
+        if self.no_std {
+            quote! {
+                use alloc::{vec, vec::Vec};
+                use fuels_core::code_gen::function_selector::resolve_fn_selector;
+                use fuels_core::types::*;
+                use fuels_core::{EnumSelector, Parameterize, Tokenizable, Token, Identity, try_from_bytes};
+                use fuels_types::enum_variants::EnumVariants;
+                use fuels_types::errors::Error as SDKError;
+                use fuels_types::param_types::ParamType;
+            }
+        } else {
+            let specific_includes = if is_script {
+                quote! {
+                    use fuels::contract::script_calls::{ScriptCallHandler, ScriptCall};
+                    use fuels::core::abi_encoder::ABIEncoder;
+                    use fuels::core::parameters::TxParameters;
+                    use std::marker::PhantomData;
+                }
+            } else {
+                quote! {
+                    use fuels::contract::contract::{
+                        Contract,
+                        ContractCallHandler,
+                        get_decoded_output
+                    };
+                    use fuels::core::abi_decoder::ABIDecoder;
+                    use fuels::core::code_gen::{
+                        function_selector::resolve_fn_selector,
+                        get_logs_hashmap,
+                    };
+                    use fuels::core::{EnumSelector, StringToken, Identity};
+                    use fuels::types::{
+                        ResolvedLog,
+                        bech32::Bech32ContractId
+                    };
+                    use std::collections::{HashSet, HashMap};
+                    use std::str::FromStr;
+                }
+            };
+            quote! {
+                use fuels::contract::logs::LogDecoder;
+                use fuels::core::types::*;
+                use fuels::core::{Tokenizable, Token, Parameterize, try_from_bytes};
+                use fuels::signers::WalletUnlocked;
+                use fuels::types::enum_variants::EnumVariants;
+                use fuels::types::errors::Error as SDKError;
+                use fuels::types::param_types::ParamType;
+                use fuels::tx::{ContractId, Address, Receipt};
+                #specific_includes
+            }
+        }
+    }
+
+    pub fn contract_functions(&self) -> Result<TokenStream, Error> {
         let tokenized_functions = self
             .abi
             .functions
             .iter()
             .map(|function| expand_function(function, &self.types))
             .collect::<Result<Vec<TokenStream>, Error>>()?;
-
         Ok(quote! { #( #tokenized_functions )* })
+    }
+
+    pub fn script_function(&self) -> Result<TokenStream, Error> {
+        let functions = self
+            .abi
+            .functions
+            .iter()
+            .filter(|function| function.name == "main")
+            .collect::<Vec<&ABIFunction>>();
+
+        if let [main_function] = functions.as_slice() {
+            let tokenized_function = generate_script_main_function(main_function, &self.types)?;
+            Ok(quote! { #tokenized_function })
+        } else {
+            Err(Error::CompilationError(
+                "The script must have one function named `main` to compile!".to_string(),
+            ))
+        }
     }
 
     fn abi_structs(&self) -> Result<TokenStream, Error> {
