+++ packages/fuels-signers/src/wallet.rs
@@ -83,7 +83,6 @@ pub struct Wallet {
 ///   Ok(())
 /// }
 /// ```
-/// [`Signature`]: fuels_core::signature::Signature
 #[derive(Clone, Debug)]
 pub struct WalletUnlocked {
     wallet: Wallet,
