+++ packages/fuels/tests/contracts.rs
@@ -302,7 +302,7 @@ async fn test_large_return_data() -> Result<(), Error> {
 
     let res = contract_methods.get_contract_id().call().await?;
 
-    // First `value` is from `CallResponse`.
+    // First `value` is from `FuelCallResponse`.
     // Second `value` is from the `ContractId` type.
     assert_eq!(
         res.value,
