+++ packages/fuels/tests/scripts/basic_script/src/main.sw
@@ -1,0 +1,11 @@
+script;
+
+fn main(a: u64, b: u32) -> str[5] {
+    if a < b {
+        let my_string: str[5] = "hello";
+        my_string
+    } else {
+        let my_string: str[5] = "heyoo";
+        my_string
+    }
+}
