+++ packages/fuels/tests/scripts/script_tuple_argument/Forc.toml
@@ -1,0 +1,7 @@
+[project]
+authors = ["Fuel Labs <contact@fuel.sh>"]
+entry = "main.sw"
+license = "Apache-2.0"
+name = "script_tuple_argument"
+
+[dependencies]
