+++ packages/fuels/tests/scripts/script_with_arguments/Forc.toml
@@ -1,0 +1,7 @@
+[project]
+authors = ["Fuel Labs <contact@fuel.sh>"]
+entry = "main.sw"
+license = "Apache-2.0"
+name = "script_with_arguments"
+
+[dependencies]
